Refactor: 코드 병합 중 파일 위치로 인한 문제 해결
# Conflicts:
#	src/main/java/com/stylemycloset/user/dto/request/UserCreateRequest.java
