--- src/main/java/com/stylemycloset/user/dto/UserCreateRequest.java
@@ -1,8 +1,0 @@
-package com.stylemycloset.user.dto;
-
-public record UserCreateRequest(
-    String name,
-    String email,
-    String password
-) {
-}
